# openapi.yaml
openapi: 3.1.0
info:
  title: OssPilot Ops API
  version: 0.1.0
paths:
...
  /api/users:
    get:
      summary: List ops users
      responses:
        "200":
          description: items and total
        "403":
          description: not platform_admin
    post:
      summary: Create ops user
      responses:
        "201":
          description: created
        "409":
          description: username taken
  /api/users/{user_id}:
    get:
      summary: Get ops user
      responses:
        "200":
          description: user
        "404":
          description: not found
    put:
      summary: Update ops user
      responses:
        "200":
          description: updated
    delete:
      summary: Delete ops user
      responses:
        "204":
          description: deleted
...
  /api/regions:
    get:
      summary: List storage regions
      responses:
        "200":
          description: items and total
    post:
      summary: Create storage region
      responses:
        "201":
          description: created
        "409":
          description: code taken
  /api/regions/{region_id}:
    put:
      summary: Update storage region
      responses:
        "200":
          description: updated
        "404":
          description: not found
    delete:
      summary: Delete storage region
      responses:
        "204":
          description: deleted
        "409":
          description: default region
...